types and models improved
