+++ docs/examples/code/crawl_website_with_relative_links_same_hostname.py
@@ -1,7 +1,7 @@
 import asyncio
 
+from crawlee import EnqueueStrategy
 from crawlee.beautifulsoup_crawler import BeautifulSoupCrawler, BeautifulSoupCrawlingContext
-from crawlee.enqueue_strategy import EnqueueStrategy
 
 
 async def main() -> None:
