+++ src/crawlee/__init__.py
@@ -1,7 +1,7 @@
 from importlib import metadata
 
-from ._autoscaling.types import ConcurrencySettings
 from ._models import Request
+from ._types import ConcurrencySettings, EnqueueStrategy
 from ._utils.globs import Glob
 
 __version__ = metadata.version('crawlee')
