+++ tests/unit/_autoscaling/test_autoscaled_pool.py
@@ -11,7 +11,8 @@ from unittest.mock import Mock
 import pytest
 
 from crawlee._autoscaling import AutoscaledPool, SystemStatus
-from crawlee._autoscaling.types import ConcurrencySettings, LoadRatioInfo, SystemInfo
+from crawlee._autoscaling.types import LoadRatioInfo, SystemInfo
+from crawlee._types import ConcurrencySettings
 from crawlee._utils.measure_time import measure_time
