+++ tests/unit/basic_crawler/test_basic_crawler.py
@@ -14,14 +14,12 @@ from unittest.mock import Mock
 import httpx
 import pytest
 
-from crawlee import Glob
-from crawlee._autoscaling.types import ConcurrencySettings
+from crawlee import ConcurrencySettings, EnqueueStrategy, Glob
 from crawlee._models import BaseRequestData, Request
 from crawlee._statistics.models import FinalStatistics
 from crawlee._types import AddRequestsKwargs, BasicCrawlingContext, HttpHeaders
 from crawlee.basic_crawler import BasicCrawler
 from crawlee.configuration import Configuration
-from crawlee.enqueue_strategy import EnqueueStrategy
 from crawlee.errors import SessionError, UserDefinedErrorHandlerError
 from crawlee.storages import Dataset, KeyValueStore, RequestList, RequestQueue
