+++ tests/unit/parsel_crawler/test_parsel_crawler.py
@@ -8,7 +8,7 @@ import pytest
 import respx
 from httpx import Response
 
-from crawlee._autoscaling.types import ConcurrencySettings
+from crawlee import ConcurrencySettings
 from crawlee._models import BaseRequestData
 from crawlee.parsel_crawler import ParselCrawler
 from crawlee.storages import RequestList
